parseAPI: return size_t from Function::num_blocks()
Changing to size_t from int better matches the size of the underlying type,
and also allows comparison to other sizes without any "signed vs. unsigned"
warnings.  This function is new to 8.2, so API/ABI isn't a concern.

A few places which used num_blocks() in dyninstAPI are also updated to
size_t, but none of these are in public headers.

+++ dyninstAPI/src/function.C
@@ -268,7 +268,7 @@ mapped_object *func_instance::obj() const { return mod()->obj(); }
 AddressSpace *func_instance::proc() const { return obj()->proc(); }
 
 const func_instance::BlockSet &func_instance::unresolvedCF() {
-   if (ifunc()->getPrevBlocksUnresolvedCF() != (int)ifunc()->num_blocks()) {
+   if (ifunc()->getPrevBlocksUnresolvedCF() != ifunc()->num_blocks()) {
        ifunc()->setPrevBlocksUnresolvedCF(ifunc()->num_blocks());
        // A block has unresolved control flow if it has an indirect
        // out-edge.
@@ -779,7 +779,7 @@ void func_instance::removeBlock(block_instance *block) {
     BlockSet::iterator bit = unresolvedCF_.find(block);
     if (bit != unresolvedCF_.end()) {
         unresolvedCF_.erase(bit);
-        int prev = ifunc()->getPrevBlocksUnresolvedCF();
+        size_t prev = ifunc()->getPrevBlocksUnresolvedCF();
         if (prev > 0) {
            ifunc()->setPrevBlocksUnresolvedCF(prev - 1);
         }
@@ -819,7 +819,7 @@ void func_instance::add_block_cb(block_instance * /*block*/)
       // these if cases will never execute anyway, at least not 
       // when we intend them to
     if (block->llb()->unresolvedCF()) {
-       int prev = ifunc()->getPrevBlocksUnresolvedCF();
+       size_t prev = ifunc()->getPrevBlocksUnresolvedCF();
        if (ifunc()->blocks().size() == prev) {
           unresolvedCF_.insert(block);
           ifunc()->setPrevBlocksUnresolvedCF(prev+1);
